Move tests to appropriate module
