+++ py-polars/tests/unit/namespaces/test_datetime.py
@@ -1118,48 +1118,3 @@ def test_agg_expr() -> None:
     )
 
     assert_frame_equal(df.select(pl.all().mean()), expected)
-
-
-@pytest.mark.parametrize(
-    ("left_time_unit", "right_time_unit", "expected_time_unit", "expected_nanoseconds"),
-    [
-        ("ms", "ms", "ms", 123_000_000 - 123_000_000),
-        ("ms", "us", "us", 123_000_000 - 123_456_000),
-        ("us", "ms", "us", 123_456_000 - 123_000_000),
-        ("us", "us", "us", 123_456_000 - 123_456_000),
-        ("ms", "ns", "ns", 123_000_000 - 123_456_789),
-        ("us", "ns", "ns", 123_456_000 - 123_456_789),
-        ("ns", "ms", "ns", 123_456_789 - 123_000_000),
-        ("ns", "us", "ns", 123_456_789 - 123_456_000),
-        ("ns", "ns", "ns", 123_456_789 - 123_456_789),
-    ],
-)
-def test_subtract_datetimes_keep_subsecond_precision(
-    left_time_unit: TimeUnit,
-    right_time_unit: TimeUnit,
-    expected_time_unit: TimeUnit,
-    expected_nanoseconds: int,
-) -> None:
-    df_datetimes = pl.DataFrame(
-        {
-            "left": ["2020-01-01T00:00:01.123456789"],
-            "right": ["2020-01-01T00:00:00.123456789"],
-        }
-    ).select(
-        pl.col("left").str.to_datetime(time_unit=left_time_unit),
-        pl.col("right").str.to_datetime(time_unit=right_time_unit),
-    )
-
-    result_df = df_datetimes.select(
-        (pl.col("left") - pl.col("right")).alias("left-right")
-    )
-
-    expected_df = pl.select(
-        pl.duration(
-            seconds=1,
-            nanoseconds=expected_nanoseconds,
-            time_unit=expected_time_unit,
-        ).alias("left-right")
-    )
-
-    assert_frame_equal(result_df, expected_df)
