+++ py-polars/tests/unit/operations/arithmetic/test_arithmetic.py
@@ -1,6 +1,8 @@
+from __future__ import annotations
+
 import operator
 from datetime import date, datetime, timedelta
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 import numpy as np
 import pytest
@@ -8,6 +10,9 @@ import pytest
 import polars as pl
 from polars.datatypes import FLOAT_DTYPES, INTEGER_DTYPES
 from polars.testing import assert_frame_equal, assert_series_equal
+
+if TYPE_CHECKING:
+    from polars.type_aliases import TimeUnit
 
 
 def test_sqrt_neg_inf() -> None:
@@ -293,3 +298,102 @@ def test_null_column_arithmetic(op: Any) -> None:
     # test broadcast left
     output_df = df.select(op(pl.Series("a", [None]), pl.col("a")))
     assert_frame_equal(expected_df, output_df)
+
+
+@pytest.mark.parametrize(
+    (
+        "dt_time_unit",
+        "duration_time_unit",
+        "expected_time_unit",
+        "expected_nanoseconds",
+    ),
+    [
+        ("ms", "ms", "ms", 124000000),
+        ("ms", "us", "us", 123001000),
+        ("us", "ms", "us", 124456000),
+        ("us", "us", "us", 123457000),
+        ("ms", "ns", "ns", 123000001),
+        ("us", "ns", "ns", 123456001),
+        ("ns", "ms", "ns", 124456789),
+        ("ns", "us", "ns", 123457789),
+        ("ns", "ns", "ns", 123456790),
+    ],
+)
+@pytest.mark.parametrize("op", ["add-left", "add-right"])
+def test_datetime_add_duration_subsecond(
+    dt_time_unit: TimeUnit,
+    duration_time_unit: TimeUnit,
+    expected_time_unit: TimeUnit,
+    op: str,
+    expected_nanoseconds: int,
+) -> None:
+    df_datetimes = pl.DataFrame({"ts": ["2020-01-01T00:00:00.123456789"]}).select(
+        pl.col("ts").str.to_datetime(time_unit=dt_time_unit)
+    )
+
+    verbose_duration_unit = {
+        "ms": "milliseconds",
+        "us": "microseconds",
+        "ns": "nanoseconds",
+    }[duration_time_unit]
+
+    duration = pl.duration(
+        **{verbose_duration_unit: 1},
+        time_unit=duration_time_unit,
+    )
+
+    if op == "add-left":
+        select_expr = duration + pl.col("ts")
+    else:
+        select_expr = pl.col("ts") + duration
+
+    result_df = df_datetimes.select(select_expr.alias("ts"))
+    expected = pl.DataFrame(
+        {"ts": [f"2020-01-01T00:00:00.{expected_nanoseconds}"]}
+    ).select(pl.col("ts").str.to_datetime(time_unit=expected_time_unit))
+    assert_frame_equal(result_df, expected)
+
+
+@pytest.mark.parametrize(
+    ("left_time_unit", "right_time_unit", "expected_time_unit", "expected_nanoseconds"),
+    [
+        ("ms", "ms", "ms", 123_000_000 - 123_000_000),
+        ("ms", "us", "us", 123_000_000 - 123_456_000),
+        ("us", "ms", "us", 123_456_000 - 123_000_000),
+        ("us", "us", "us", 123_456_000 - 123_456_000),
+        ("ms", "ns", "ns", 123_000_000 - 123_456_789),
+        ("us", "ns", "ns", 123_456_000 - 123_456_789),
+        ("ns", "ms", "ns", 123_456_789 - 123_000_000),
+        ("ns", "us", "ns", 123_456_789 - 123_456_000),
+        ("ns", "ns", "ns", 123_456_789 - 123_456_789),
+    ],
+)
+def test_subtract_datetimes_keep_subsecond_precision(
+    left_time_unit: TimeUnit,
+    right_time_unit: TimeUnit,
+    expected_time_unit: TimeUnit,
+    expected_nanoseconds: int,
+) -> None:
+    df_datetimes = pl.DataFrame(
+        {
+            "left": ["2020-01-01T00:00:01.123456789"],
+            "right": ["2020-01-01T00:00:00.123456789"],
+        }
+    ).select(
+        pl.col("left").str.to_datetime(time_unit=left_time_unit),
+        pl.col("right").str.to_datetime(time_unit=right_time_unit),
+    )
+
+    result_df = df_datetimes.select(
+        (pl.col("left") - pl.col("right")).alias("left-right")
+    )
+
+    expected_df = pl.select(
+        pl.duration(
+            seconds=1,
+            nanoseconds=expected_nanoseconds,
+            time_unit=expected_time_unit,
+        ).alias("left-right")
+    )
+
+    assert_frame_equal(result_df, expected_df)
